move f_table into python binding
