+++ src/lib.rs
@@ -4,7 +4,6 @@ mod core_ops;
 pub mod cost;
 mod dialect;
 mod err_utils;
-pub mod f_table;
 mod gen;
 mod int_to_bytes;
 pub mod more_ops;
