+++ src/py/mod.rs
@@ -1,5 +1,6 @@
 mod adapt_response;
 mod api;
+mod f_table;
 mod lazy_node;
 mod run_generator;
 mod run_program;
